Update README.md and Worked Towards Revising Barometer Return Type
Changed the Return Type to Unsigned Long by Sending Through a Simple Rounding Function for Longs
Changed the README.md to reflect succesful testing of the A4 and A5 connection even through the Trackuino headers

+++ trackuino/sensors_avr.h
@@ -21,7 +21,7 @@
 
 void sensors_setup();
 #ifdef USE_BAROMETER
-  double sensors_barometer_pressure();
+  unsigned long sensors_barometer_pressure();
 #endif
 unsigned long sensors_aref();
 long sensors_internal_temp();
